requirements.txt
fastparquet==2024.2.0
boto3==1.34.162
s3fs==2024.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1

pydantic==2.9.2
pydantic-settings==2.2.1
kagglehub

apache-airflow==2.9.3
apache-airflow-providers-apache-spark==4.8.0
apache-airflow-providers-http==4.10.0

pyspark==4.0.0
apache-airflow-providers-postgres==5.10.0